refactor(deps): extract bare-cache helpers to bare_cache module to clear file-length guardrail
Splits github_downloader.py (was 2751 lines, exceeded 2400-line CI guardrail)
by moving bare-cache helpers and the clone-failure error-message builder
into a new src/apm_cli/deps/bare_cache.py module:

  - _scrub_bare_remote_url
  - bare_clone_with_fallback
  - materialize_from_bare
  - clone_with_fallback (parameterized via repo_cls for test patchability)
  - build_clone_failure_message

GitHubPackageDownloader retains thin delegating instance methods
(_clone_with_fallback, _bare_clone_with_fallback, _materialize_from_bare)
so existing patch.object call sites continue to work unchanged.

Existing tests that patch 'apm_cli.deps.github_downloader.Repo' continue
to function because the thin wrappers resolve Repo from the
github_downloader module namespace at call time and pass it as repo_cls.

github_downloader.py: 2751 -> 2383 lines.
